Score loop counts using enrichment of foreground over background

Given a hi-c file or InteractionArray, pseudocounts are added to raw counts and the median of the foreground over the background is added as a score for each given interaction.

Usage

# S4 method for class 'GInteractions,character'
scoreInteractions(
  x,
  hicFile,
  loopCalls,
  fgSize = 0,
  bg = "bowtie",
  bgGap = 0,
  bgSize = 5,
  pseudo,
  pruneUnder = 0,
  truncate = F,
  resolution,
  norm = "KR"
)

Arguments

x

GInteractions object or InteractionArray object containing interactions of interest See also `mariner::pixelsToMatrices`

loopCalls

GInteractions object containing "truth" loop calls for optimizing pseudocounts See also `qc_scoreInteractions`

fgSize

manhattan distance surrounding center pixel to include in foreground selection See also parameter `n` in `mariner::selectRadius`

bg

character keyword for background selection shape One of `c("bowtie", "donut")`

bgGap

number of pixels between foreground and background shapes

bgSize

Integer describing the width (in pixels) of the background shape

pseudo

number of pseudo counts to add to raw Hi-C counts

pruneUnder

numeric, all scores under this value will be set to 0

truncate

logical indicating whether to set any values above the 99th percentile to the 99th percentile value

resolution

Integer (numeric) describing the resolution (range widths) of the paired data. Used only if GInteractions object is supplied for x. By default, will match the width of regions in `x`

norm

String (length one character vector) describing the Hi-C normalization to apply. Use `strawr::readHicNormTypes()` to see accepted values for each file in `files`. Required only if GInteractions object is supplied for x.

file

Character file paths to a `.hic` file. Required only if GInteractions object is supplied for x.

interactions

GInteraction object of interactions from Hi-C data

Examples

## add examples here